# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: NexoShop API
  version: 1.0.0
  description: >-
    Contrato estático de la API REST académica de NexoShop. Describe el backend
    monolítico implementado para usuarios, catálogo, carritos y pedidos. No
    incluye autenticación, JWT, autorización ni pagos externos.
servers:
  - url: http://localhost:8080
    description: Servidor local
...
paths:
...
  /api/v1/categories:
    post:
      tags: [Categories]
      summary: Crear categoría
      description: Crea una categoría activa.
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
            examples:
              default:
                $ref: '#/components/examples/CategoryRequestExample'
      responses:
        '201':
          description: Categoría creada.
          headers:
            Location:
              description: URI de la categoría creada.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Categories]
      summary: Listar categorías
      description: Devuelve categorías ordenadas por nombre e identificador. activeOnly es opcional y filtra únicamente categorías activas cuando vale true.
      operationId: listCategories
      parameters:
        - $ref: '#/components/parameters/ActiveOnly'
      responses:
        '200':
          description: Lista de categorías.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/categories/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    put:
      tags: [Categories]
      summary: Actualizar categoría
      description: Actualiza nombre y descripción de una categoría.
      operationId: updateCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRequest'
      responses:
        '200':
          description: Categoría actualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Categories]
      summary: Consultar categoría
      description: Devuelve una categoría por identificador.
      operationId: getCategory
      responses:
        '200':
          description: Categoría encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/categories/{id}/activate:
    post:
      tags: [Categories]
      summary: Activar categoría
      description: Marca la categoría como activa.
      operationId: activateCategory
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Categoría activada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/categories/{id}/deactivate:
    post:
      tags: [Categories]
      summary: Desactivar categoría
      description: Marca la categoría como inactiva.
      operationId: deactivateCategory
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Categoría desactivada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/products:
    post:
      tags: [Products]
      summary: Crear producto
      description: Crea un producto dentro de una categoría activa.
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
            examples:
              default:
                $ref: '#/components/examples/ProductRequestExample'
      responses:
        '201':
          description: Producto creado.
          headers:
            Location:
              description: URI del producto creado.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Products]
      summary: Listar productos
      description: Devuelve una página estable de productos ordenada por id ascendente. activeOnly filtra productos activos y calcula la metadata sobre ese subconjunto.
      operationId: listProducts
      parameters:
        - $ref: '#/components/parameters/ActiveOnly'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Página de productos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/products/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    put:
      tags: [Products]
      summary: Actualizar producto
      description: Actualiza los datos del producto sin cambiar su stock.
      operationId: updateProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Producto actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Products]
      summary: Consultar producto
      description: Devuelve un producto por identificador.
      operationId: getProduct
      responses:
        '200':
          description: Producto encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/products/sku/{sku}:
    get:
      tags: [Products]
      summary: Consultar producto por SKU
      description: Normaliza el SKU y devuelve el producto coincidente.
      operationId: getProductBySku
      parameters:
        - $ref: '#/components/parameters/Sku'
      responses:
        '200':
          description: Producto encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/products/{id}/activate:
    post:
      tags: [Products]
      summary: Activar producto
      description: Activa el producto si su categoría está activa.
      operationId: activateProduct
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Producto activado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/products/{id}/deactivate:
    post:
      tags: [Products]
      summary: Desactivar producto
      description: Desactiva el producto.
      operationId: deactivateProduct
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Producto desactivado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/products/{id}/stock/increase:
    post:
      tags: [Products]
      summary: Incrementar stock
      description: Incrementa el stock con una cantidad positiva.
      operationId: increaseProductStock
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockAdjustmentRequest'
            examples:
              default:
                $ref: '#/components/examples/StockAdjustmentExample'
      responses:
        '200':
          description: Stock incrementado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/products/{id}/stock/decrease:
    post:
      tags: [Products]
      summary: Decrementar stock
      description: Reduce el stock sin permitir valores negativos.
      operationId: decreaseProductStock
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockAdjustmentRequest'
            examples:
              default:
                $ref: '#/components/examples/StockAdjustmentExample'
      responses:
        '200':
          description: Stock decrementado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: Identificador positivo de 64 bits.
      schema:
        type: integer
        format: int64
        minimum: 1
...
    Sku:
      name: sku
      in: path
      required: true
      description: SKU textual del producto.
      schema:
        type: string
    Page:
      name: page
      in: query
      required: false
      description: Página basada en cero.
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      description: Cantidad positiva de elementos por página.
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 20
    ActiveOnly:
      name: activeOnly
      in: query
      required: false
      description: Cuando es true, devuelve únicamente recursos activos. Ausente o false conserva el listado completo.
      schema:
        type: boolean
        default: false
  responses:
    BadRequest:
      description: Solicitud inválida; puede ser VALIDATION_ERROR o MALFORMED_REQUEST.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation:
              $ref: '#/components/examples/ValidationErrorExample'
            malformed:
              $ref: '#/components/examples/MalformedErrorExample'
    NotFound:
      description: Recurso inexistente (RESOURCE_NOT_FOUND).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              $ref: '#/components/examples/NotFoundErrorExample'
    Conflict:
      description: Conflicto de duplicidad, regla de negocio o integridad (409).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            duplicate:
              $ref: '#/components/examples/DuplicateErrorExample'
            business:
              $ref: '#/components/examples/BusinessErrorExample'
...
    InternalError:
      description: Error inesperado sin detalles internos (INTERNAL_ERROR).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              $ref: '#/components/examples/InternalErrorExample'
  schemas:
...
    CreateCategoryRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
        description:
          type: string
          maxLength: 500
          nullable: true
    UpdateCategoryRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
        description:
          type: string
          maxLength: 500
          nullable: true
    CreateProductRequest:
      type: object
      required: [categoryId, sku, name, price, stock]
      properties:
        categoryId:
          type: integer
          format: int64
          minimum: 1
        sku:
          type: string
          maxLength: 50
          minLength: 1
        name:
          type: string
          maxLength: 150
          minLength: 1
        description:
          type: string
          nullable: true
        price:
          $ref: '#/components/schemas/PositiveMoney'
        stock:
          type: integer
          format: int32
          minimum: 0
        imageUrl:
          type: string
          maxLength: 500
          nullable: true
    UpdateProductRequest:
      type: object
      required: [categoryId, sku, name, price]
      properties:
        categoryId:
          type: integer
          format: int64
          minimum: 1
        sku:
          type: string
          maxLength: 50
          minLength: 1
        name:
          type: string
          maxLength: 150
          minLength: 1
        description:
          type: string
          nullable: true
        price:
          $ref: '#/components/schemas/PositiveMoney'
        imageUrl:
          type: string
          maxLength: 500
          nullable: true
    StockAdjustmentRequest:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          format: int32
          minimum: 1
...
    CategoryResponse:
      type: object
      required: [id, name, active, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
          nullable: true
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProductResponse:
      type: object
      required: [id, categoryId, categoryName, sku, name, price, stock, active, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: string
        sku:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          $ref: '#/components/schemas/PositiveMoney'
        stock:
          type: integer
          format: int32
          minimum: 0
        imageUrl:
          type: string
          nullable: true
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ProductPageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages, first, last]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
...
    FieldErrorResponse:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      required: [timestamp, status, code, message, path, fieldErrors]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        code:
          type: string
          enum: [VALIDATION_ERROR, MALFORMED_REQUEST, RESOURCE_NOT_FOUND, DUPLICATE_RESOURCE, BUSINESS_RULE, INTEGRITY_VIOLATION, INTERNAL_ERROR]
        message:
          type: string
        path:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorResponse'
    PositiveMoney:
      type: number
      format: double
      minimum: 0
      exclusiveMinimum: true
      maximum: 9999999999.99
      multipleOf: 0.01
      example: 49.90
...
  examples:
...
    CategoryRequestExample:
      summary: Categoría ficticia
      value:
        name: Libros
        description: Lecturas de ejemplo
    ProductRequestExample:
      summary: Producto ficticio
      value:
        categoryId: 2
        sku: BOOK-DEMO-01
        name: Libro de ejemplo
        description: Producto ficticio para documentación
        price: 49.90
        stock: 12
        imageUrl: https://example.test/book.jpg
    StockAdjustmentExample:
      summary: Ajuste de stock ficticio
      value:
        quantity: 3
...
    ValidationErrorExample:
      summary: Error de validación
      value:
        timestamp: '2026-08-20T05:27:46Z'
        status: 400
        code: VALIDATION_ERROR
        message: Request validation failed
        path: /api/v1/products
        fieldErrors:
          - field: price
            message: must be greater than 0.00
    MalformedErrorExample:
      summary: Parámetro inválido
      value:
        timestamp: '2026-08-20T05:27:46Z'
        status: 400
        code: MALFORMED_REQUEST
        message: Request parameter has an invalid format
        path: /api/v1/categories/abc
        fieldErrors: []
    NotFoundErrorExample:
      summary: Recurso inexistente
      value:
        timestamp: '2026-08-20T05:27:46Z'
        status: 404
        code: RESOURCE_NOT_FOUND
        message: 'Product not found: 3'
        path: /api/v1/products/3
        fieldErrors: []
    DuplicateErrorExample:
      summary: Recurso duplicado
      value:
        timestamp: '2026-08-20T05:27:46Z'
        status: 409
        code: DUPLICATE_RESOURCE
        message: Email already exists
        path: /api/v1/users
        fieldErrors: []
    BusinessErrorExample:
      summary: Regla de negocio
      value:
        timestamp: '2026-08-20T05:27:46Z'
        status: 409
        code: BUSINESS_RULE
        message: Insufficient stock
        path: /api/v1/users/2/orders/checkout
        fieldErrors: []
    InternalErrorExample:
      summary: Error interno
      value:
        timestamp: '2026-08-20T05:27:46Z'
        status: 500
        code: INTERNAL_ERROR
        message: An unexpected error occurred
        path: /api/v1/products
        fieldErrors: []